html/infrastructure/urls/terminology-0/document-base-url.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: document base URL</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/infrastructure.html#document-base-url">
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/infrastructure.html#terminology-0">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  iframe { display: none }
</style>
<body onload="on_load()">
  <div id="log"></div>
  <script>

  var t1 = async_test("The document base URL of a document containing one or more base elements with href attributes is the frozen base URL of the first base element in the document that has an href attribute, in tree order."),
      t2 = async_test("The fallback base URL of a document containing no base element is the document's address."),
      t3 = async_test("The fallback base URL of a document whose address is about:blank is the document base URL of the creator document.");
      t4 = async_test("The fallback base URL of an iframe srcdoc document is the document base URL of the document's browsing context's browsing context container's document.");

  function on_load() {
    t1.step(function () {
      var base = document.createElement("base");
      base.setAttribute("href", "/foo/bar");
      document.head.appendChild(base);
      assert_equals(document.baseURI, base.href, "The document base URL should be URL of the first base element that has an href attribute.");
    });
    t1.done();

    t2.step(function () {
      var fr = document.createElement("iframe"),
          iframe_doc;
      fr.setAttribute("src", "/common/blank.html");
      document.body.appendChild(fr);
      iframe_doc = document.getElementsByTagName("iframe")[0].contentDocument;
      iframe_doc.addEventListener("load", function () {
        assert_equals(iframe_doc.baseURI, iframe_doc.location, "The document base URL should be the document's address.");
      });
    });
    t2.done();

    t3.step(function () {
      var fr = document.createElement("iframe"),
          iframe_doc;
      fr.setAttribute("src", "about:blank");
      document.body.appendChild(fr);
      iframe_doc = document.getElementsByTagName("iframe")[1].contentDocument;
      iframe_doc.addEventListener("load", function () {
        assert_equals(iframe_doc.baseURI, document.baseURI, "The document base URL should be the creator document's base URL.");
      });
    });
    t3.done();

    t4.step(function () {
      var fr = document.createElement("iframe"),
          iframe_doc;
      fr.setAttribute("srcdoc", "<p>foobar</p>");
      document.body.appendChild(fr);
      iframe_doc = document.getElementsByTagName("iframe")[2].contentDocument;
      iframe_doc.addEventListener("load", function () {
        assert_equals(iframe_doc.baseURI, document.baseURI, "The document base URL should be the containing document's base URL.");
      });
    });
    t4.done();
  }
  </script>
</body>
